Remove API products
